Trait mina_consensus::common::Chain

pub trait Chain<T> where
    T: Header, {
    fn push(&mut self, new: T) -> Result<(), ConsensusError>;
    fn top(&self) -> Option<&T>;
    fn consensus_state(&self) -> Option<&ConsensusState>;
    fn global_slot(&self) -> Option<&GlobalSlot>;
    fn epoch_slot(&self) -> Option<u32>;
    fn length(&self) -> usize;
    fn last_vrf_hash(&self) -> Option<Fp>;
    fn state_hash(&self) -> Option<Fp>;
    fn genesis_block(&self) -> Option<&ProtocolState>;
}

Trait that represents a chain of block data structure

Required Methods

fn push(&mut self, new: T) -> Result<(), ConsensusError>

Pushes an item into the chain

fn top(&self) -> Option<&T>

This function returns the last block of a given chain. The input is a chain C and the output is last block of C (i.e. the block with greatest height).

fn consensus_state(&self) -> Option<&ConsensusState>

he function returns the consensus state of a block or chain. The input is a block or chain X and the output is the consensus state.

fn global_slot(&self) -> Option<&GlobalSlot>

The function returns the global slot number of a chain or block. The input X is either a chain or block and the output is the global slot number.

fn epoch_slot(&self) -> Option<u32>

The function computes the epoch slot number of a block. The output is the epoch slot number in [0, slots_per_epoch].

fn length(&self) -> usize

The function the length of a chain. The output is the length of the chain in blocks.

fn last_vrf_hash(&self) -> Option<Fp>

This function returns the hex digest of the hash of the last VRF output of a given chain. The input is a chain C and the output is the hash digest.

fn state_hash(&self) -> Option<Fp>

This function returns hash of the top block’s protocol state for a given chain. The input is a chain C and the output is the hash.

fn genesis_block(&self) -> Option<&ProtocolState>

Gets ProtocolState of the genesis block