Trait mina_consensus::common::Consensus

pub trait Consensus {
    type Chain;

    fn select_secure_chain<'a>(
        &'a self,
        candidates: &'a [Self::Chain]
    ) -> Result<&'a ProtocolStateChain, ConsensusError>;
    fn select_longer_chain<'a>(
        &'a self,
        candidate: &'a ProtocolStateChain
    ) -> Result<&'a ProtocolStateChain, ConsensusError>;
    fn is_short_range(
        &self,
        candidate: &ProtocolStateChain
    ) -> Result<bool, ConsensusError>;
    fn relative_min_window_density(
        &self,
        candidate: &ProtocolStateChain
    ) -> Result<u32, ConsensusError>;
    fn config(&self) -> ConsensusConstants;
}

A trait that defines operations for consensus calculation

Required Associated Types

type Chain

Chain type

Required Methods

fn select_secure_chain<'a>(
&'a self,
candidates: &'a [Self::Chain]
) -> Result<&'a ProtocolStateChain, ConsensusError>

Top level API to select between chains during a fork.

fn select_longer_chain<'a>(
&'a self,
candidate: &'a ProtocolStateChain
) -> Result<&'a ProtocolStateChain, ConsensusError>

Selects the longer chain when there’s a short range fork.

fn is_short_range(
&self,
candidate: &ProtocolStateChain
) -> Result<bool, ConsensusError>

Checks whether the fork is short range wrt to candidate chain

fn relative_min_window_density(
&self,
candidate: &ProtocolStateChain
) -> Result<u32, ConsensusError>

Calculates the relate minimum window density wrt to candidate chain.

fn config(&self) -> ConsensusConstants

Constants used for consensus